requirements.txt
crewai
litellm
streamlit>=1.20
python-dotenv
